+++ fotoskazka/app/Filament/Resources/Albums/Pages/UploadPhotos.php
@@ -48,6 +48,7 @@ class UploadPhotos extends Page
                                 'homepage' => 'Главная страница',
                                 'service' => 'Услуга',
                                 'client' => 'Клиентская галерея',
+                                'behind_the_scenes' => 'Фото со съёмок',
                             ])
                             ->live(),
                         Textarea::make('description')